+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/common/priv/CommonPrivateChatsController.java
@@ -41,7 +41,7 @@ public final class CommonPrivateChatsController extends PrivateChatsController {
     public CommonPrivateChatsView createAndGetView() {
         return new CommonPrivateChatsView((CommonPrivateChatsModel) model,
                 this,
-                chatMessagesComponent.getRoot(),
+                chatMessageContainerController.getView().getRoot(),
                 channelSidebar.getRoot());
     }
 }
